% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bymans_functions.R
\name{dygraph_modUI}
\alias{dygraph_modUI}
\title{This was an attempt to modularize Bymans functions. UI module to be used with "dygraph_mod". May not work...}
\usage{
dygraph_modUI(id)
}
\arguments{
\item{id}{}
}
\description{
This was an attempt to modularize Bymans functions. UI module to be used with "dygraph_mod". May not work...
}
