% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Function for single station and single model plots. Used in the server module "forecast_plot_mod"}
\usage{
forecast_plot(OBS, dat)
}
\arguments{
\item{dat}{}
}
\description{
Function for single station and single model plots. Used in the server module "forecast_plot_mod"
}
\examples{
In "forecast_plot_mod"
output$plot <- renderPlotly(forecast_plot(subset_OBS(), subset2plot())
}
