% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{forecast_plot_mod}
\alias{forecast_plot_mod}
\title{forecast_plot_mod}
\usage{
forecast_plot_mod(input, output, session, map_input, dat)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{map_input}{}

\item{dat}{}
}
\description{
Shiny server module to plot ...
}

