% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{forecast_plot_mod_shading}
\alias{forecast_plot_mod_shading}
\title{forecast_plot_mod_shading}
\usage{
forecast_plot_mod_shading(input, output, session, map_input, dat)
}
\arguments{
\item{id}{}
}
\description{
Shiny server module to plot ...
Same plot as forecast_plot_mod but without plotly to get the shading for the current day
}

