% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{forecast_plot_mod_shadingUI}
\alias{forecast_plot_mod_shadingUI}
\title{forecast_plot_mod_shadingUI}
\usage{
forecast_plot_mod_shadingUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with "forecast_plot_mod_shading"
}

