% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{forecast_plot_shading}
\alias{forecast_plot_shading}
\title{forecast_plot_shading}
\usage{
forecast_plot_shading(dat)
}
\arguments{
\item{dat}{}
}
\description{
This function only produces "static" ggplot for the moment, but I need to find a way to do the shading with plotly
}
\examples{
forecast_plot_shading(Mistake, Mist)
}

