% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule}
\alias{mapModule}
\title{Shiny server module to be used with the UI module "mapModuleUI"}
\usage{
mapModule(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
input This is a server module which returns the "input" from the selection buttonsto the UI module "mapModuleUI"
}
\description{
Shiny server module to be used with the UI module "mapModuleUI"
}
\examples{
See server.R and UI.R for its implementation into the shiny App
}
