% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule_polygonFeature}
\alias{mapModule_polygonFeature}
\title{mapModule_polygonFeature}
\usage{
mapModule_polygonFeature(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Shiny server module to map... TO tidy up with mapModule_polygonFeature
}

