% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule_polygonFeatureUI}
\alias{mapModule_polygonFeatureUI}
\title{mapModule_polygonFeatureUI}
\usage{
mapModule_polygonFeatureUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with "mapModule_polygonFeature" ...
}

