% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModuleUI}
\alias{mapModuleUI}
\title{mapModuleUI}
\usage{
mapModuleUI(id, multiple_choice = FALSE)
}
\arguments{
\item{id}{}

\item{multiple_choice}{}
}
\description{
Shiny UI module to be used with "mapModule" ...
}

