% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{multimod_forecast_plot}
\alias{multimod_forecast_plot}
\title{multimod_forecast_plot}
\usage{
multimod_forecast_plot(obs_data = NULL, dat_1 = NULL, dat_2 = NULL,
  dat_3 = NULL, dat_4 = NULL, return_levels = NULL, gg_plot = FALSE)
}
\arguments{
\item{dat_1}{}

\item{dat_2}{}

\item{dat_3}{}

\item{dat_4}{}

\item{return_levels}{}

\item{gg_plot}{}
}
\description{
multimod_forecast_plot
}

