% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R, R/plotting_functions.R
\name{multimod_forecast_plot_EXP}
\alias{multimod_forecast_plot_EXP}
\title{multimod_forecast_plot_EXP}
\usage{
multimod_forecast_plot_EXP(dat_1 = NULL, dat_2 = NULL, dat_3 = NULL,
  dat_4 = NULL)

multimod_forecast_plot_EXP(dat_1 = NULL, dat_2 = NULL, dat_3 = NULL,
  dat_4 = NULL)
}
\arguments{
\item{dat_1}{}

\item{dat_2}{}

\item{dat_3}{}

\item{dat_4}{}

\item{input}{}

\item{output}{}

\item{session}{}

\item{selected_stations}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}
}
\description{
Shiny server module to do multi-model plots. Needs tidy up with other functions...

To tidy up probable double up with multimod_forecast_plot
}

