% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{multimod_forecast_plot_mod}
\alias{multimod_forecast_plot_mod}
\title{Shiny server module for the multimodel and multistation plot. Used in server.R and in "mapModule_polygonFeature"}
\usage{
multimod_forecast_plot_mod(input, output, session, map_input, OBS, model_1,
  model_2, model_3, model_4, return_levels = NULL)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{map_input}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}

\item{return_levels}{}
}
\description{
Shiny server module for the multimodel and multistation plot. Used in server.R and in "mapModule_polygonFeature"
}
\examples{
In server.R
 input4multi_forecast_plot <- callModule(mapModule,"multistation_map")
 callModule(multimod_forecast_plot_mod,"multistation_plot", input4multi_forecast_plot, OBS, HBV_2014, HBV_2016, DDD, HBV_past_year, flomtabell)
}
