% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{multimod_forecast_plot_modUI}
\alias{multimod_forecast_plot_modUI}
\title{multimod_forecast_plot_modUI}
\usage{
multimod_forecast_plot_modUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with multimod_forecast_plot
}

