% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{multiple_station_map}
\alias{multiple_station_map}
\title{Mapping function used by the server module "mapModule_polygonFeature"}
\usage{
multiple_station_map(stations, selected_regine_main = NULL,
  selected_name = NULL, selected_long = NULL, selected_lat = NULL,
  single_poly = FALSE, variable2plot = "Ingen", radius_function = TRUE,
  popups = FALSE)
}
\arguments{
\item{stations}{It is a global variable in the Shiny App but is included here as a parameter for general use. 
See global.R for the data preparation}

\item{selected_regine_main}{}

\item{selected_name}{}

\item{selected_long}{}

\item{selected_lat}{}

\item{single_poly}{}

\item{variable2plot}{}

\item{radius_function}{}

\item{popups}{}
}
\value{
map Returns a leaflet map to the "mapModule_polygonFeature" server module
}
\description{
Mapping function used by the server module "mapModule_polygonFeature"
}
\examples{
See use in the "mapModule_polygonFeature" module
output$map <- renderLeaflet(multiple_station_map(stations, single_poly = FALSE, variable2plot = input$variable, popups = input$popups))
}
