% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{my.color.func}
\alias{my.color.func}
\title{Function that controls the colour of the colour markers}
\usage{
my.color.func(x2plot, my.colors)
}
\arguments{
\item{x2plot}{}

\item{my.colors}{}
}
\description{
Function that controls the colour of the colour markers
}
\examples{
my.colors <- c("blue", "green", "yellow", "orange", "red", "black")
my.color.func(OK_stations$flood_warning, my.colors)
}
