% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{my.radius.func}
\alias{my.radius.func}
\title{Function that control the size of the colour markers}
\usage{
my.radius.func(x2plot, radius_function)
}
\arguments{
\item{x2plot}{}

\item{radius_function}{}
}
\description{
Function that control the size of the colour markers
}
\examples{
map <- addCircleMarkers(map, data = OK_stations, lng = ~ longitude, lat = ~ latitude, 
                 radius = ~my.radius.func(OK_stations$flood_warning, radius_function), 
                 color = ~my.color.func(OK_stations$flood_warning, my.colors), 
                 stroke = FALSE, fillOpacity = 1,
                 group = "Warning ratio available")
}
