% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{mydygraphModule}
\alias{mydygraphModule}
\title{mydygraphModule}
\usage{
mydygraphModule(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
This was an attempt to modularize Bymans functions. Server module. May not work...
}

