% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{mydygraphModuleUI}
\alias{mydygraphModuleUI}
\title{mydygraphModuleUI}
\usage{
mydygraphModuleUI(id)
}
\arguments{
\item{id}{}
}
\description{
This was an attempt to modularize Bymans functions. UI module to be used with "mydygraphModule". May not work...
}

