% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{NEW_mapModule_polygonFeatureUI}
\alias{NEW_mapModule_polygonFeatureUI}
\title{NEW_mapModule_polygonFeatureUI}
\usage{
NEW_mapModule_polygonFeatureUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with "NEW_mapModule_polygonFeature"
}

