% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{OLD_mapModule_polygonFeatureUI}
\alias{OLD_mapModule_polygonFeatureUI}
\title{OLD_mapModule_polygonFeatureUI}
\usage{
OLD_mapModule_polygonFeatureUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with "OLD_mapModule_polygonFeature"
}

