% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{OLD_multimod_forecast_plot}
\alias{OLD_multimod_forecast_plot}
\title{OLD_multimod_forecast_plot}
\usage{
OLD_multimod_forecast_plot(input, output, session, selected_stations = NULL,
  model_1, model_2, model_3, model_4, return_levels = NULL, variable_1,
  variable_2, variable_3, variable_4, type_rl)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{selected_stations}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}

\item{return_levels}{}

\item{variable_1}{}

\item{variable_2}{}

\item{variable_3}{}

\item{variable_4}{}

\item{type_rl}{}
}
\description{
Shiny server module to do multi-model plots. Needs tidy up with previous function
}

