% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{single_station_map}
\alias{single_station_map}
\title{single_station_map}
\usage{
single_station_map(stations, selected_nbname = NULL, selected_long = NULL,
  selected_lat = NULL, variable2plot = "Ingen",
  map_layer = "Openstreetmap", catchments = FALSE,
  colored_markers = FALSE, radius_function = TRUE, popups = FALSE)
}
\arguments{
\item{stations}{}

\item{selected_nbname}{}

\item{selected_long}{}

\item{selected_lat}{}

\item{variable2plot}{}

\item{map_layer}{}

\item{catchments}{}

\item{colored_markers}{}

\item{radius_function}{}

\item{popups}{}
}
\description{
mapping function
}

