% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{station_map}
\alias{station_map}
\title{Mapping function for the first tab of the flomvarsling ShinyApp. 
This function is called by the shiny server module "mapModule"}
\usage{
station_map(stations, selected_nbname = NULL, selected_long = NULL,
  selected_lat = NULL, variable2plot = "Ingen",
  map_layer = "Openstreetmap", catchments = FALSE, radius_function = TRUE,
  popups = FALSE)
}
\arguments{
\item{stations}{It is a global variable in the Shiny App but is included here as a parameter for general use. 
See global.R for the data preparation}

\item{selected_nbname}{Selected station in the following format. "2.11-narsjø". This can be a vector}

\item{selected_long}{}

\item{selected_lat}{}

\item{variable2plot}{"Fare for flom", "Usikkerhet" or "Ingen"}

\item{map_layer}{"Openstreetmap" by default. Other choices: "Topo kart" or "Flyfoto"}

\item{catchments}{Specify TRUE if you want to see the catchment boundaries}

\item{radius_function}{Default "TRUE". Functions which makes marker size proportional to the variable2plot value}

\item{popups}{Specify TRUE if you want to see have popups for the selected stations}
}
\value{
map Returns a map that is used by "mapModule"
}
\description{
Mapping function for the first tab of the flomvarsling ShinyApp. 
This function is called by the shiny server module "mapModule"
}
\examples{
This is how we use this function in the shiny server module "mapModule"
output$map <- renderLeaflet({station_map(stations, input$station, selected_long(),
selected_lat(), variable2plot = input$variable, map_layer = input$map_layer, catchments = input$catchments, popups = input$popups)})
}
