% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{SUPERCEDED_mapModule_polygonFeature}
\alias{SUPERCEDED_mapModule_polygonFeature}
\title{SUPERCEDED_mapModule_polygonFeature}
\usage{
SUPERCEDED_mapModule_polygonFeature(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Shiny server module to map ...
}

