% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{SUPERCEDED_mapModule_polygonFeatureUI}
\alias{SUPERCEDED_mapModule_polygonFeatureUI}
\title{SUPERCEDED_mapModule_polygonFeatureUI}
\usage{
SUPERCEDED_mapModule_polygonFeatureUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with "mapModule_polygonFeature" ...
}

