% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_modules.R
\name{table_mod}
\alias{table_mod}
\title{Shiny server module to render data tables}
\usage{
table_mod(input, output, session, dat)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{dat}{}
}
\description{
Shiny server module to render data tables
}
