% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_modules.R
\name{table_modUI}
\alias{table_modUI}
\title{Shiny UI module to be used with table_mod}
\usage{
table_modUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to be used with table_mod
}
\examples{
In UI.R
tabPanel("Metadata",
table_modUI("metadata_table")),
}
