% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bymans_functions.R
\name{taylor_mod}
\alias{taylor_mod}
\title{This was an attempt to modularize Bymans functions. Server module. May not work...}
\usage{
taylor_mod(input, output, session, selected_stations, model_1, model_2, model_3,
  model_4 = NULL)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{selected_stations}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}
}
\description{
This was an attempt to modularize Bymans functions. Server module. May not work...
}
