% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bymans_functions.R
\name{taylor_modUI}
\alias{taylor_modUI}
\title{This was an attempt to modularize Bymans functions. UI module to be used with "taylor_mod". May not work...}
\usage{
taylor_modUI(id)
}
\arguments{
\item{id}{}
}
\description{
This was an attempt to modularize Bymans functions. UI module to be used with "taylor_mod". May not work...
}
