% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{forecast_plot_mod}
\alias{forecast_plot_mod}
\title{Shiny server module for the single model and single station plot. Used in server.R}
\usage{
forecast_plot_mod(input, output, session, map_input, dat)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{map_input}{}

\item{dat}{}
}
\description{
Shiny server module for the single model and single station plot. Used in server.R
}
\examples{
In server.R
callModule(forecast_plot_mod,"forecast_plot_HBV_2014", input4plot_HBV_2014, HBV_2014)
}
