% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule}
\alias{mapModule}
\title{mapModule}
\usage{
mapModule(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Shiny server module to map ...
}

