% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule_polygonFeature}
\alias{mapModule_polygonFeature}
\title{Shiny server module to be used with the UI module "mapModule_polygonFeatureUI"}
\usage{
mapModule_polygonFeature(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
NULL This is a server module and could return "input" as with "mapModule", but I some problems implementing it in a similar way.
So the plotting is integrated within this module with "poly_multimod_forecast_plot_mod" which makes things a bit confusing.
This implementation could probably be simplified to resemble more "mapModule".
}
\description{
Shiny server module to be used with the UI module "mapModule_polygonFeatureUI"
}
\examples{
See server.R and UI.R for its implementation into the shiny App
}
