% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_modules.R
\name{mapModule_polygonFeatureUI}
\alias{mapModule_polygonFeatureUI}
\title{Shiny UI module to be used with the server module "mapModule_polygonFeature"}
\usage{
mapModule_polygonFeatureUI(id)
}
\arguments{
\item{id}{}
}
\value{
NULL This is a UI module which returns a map, a selection buttons: station name, map layer, catchment boundaries,
pops ups for selected stations and a choice of color markers.
}
\description{
Shiny UI module to be used with the server module "mapModule_polygonFeature"
}
\examples{
See server.R and UI.R for its implementation into the shiny App
}
