% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{multimod_forecast_plot_EXP}
\alias{multimod_forecast_plot_EXP}
\title{multimod_forecast_plot_EXP}
\usage{
multimod_forecast_plot_EXP(input, output, session, selected_stations, model_1,
  model_2, model_3, model_4 = NULL)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{selected_stations}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}
}
\description{
Shiny server module to do multi-model plots. Needs tidy up with other functions...
}

