% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{multimod_forecast_plot_mod}
\alias{multimod_forecast_plot_mod}
\title{multimod_forecast_plot_mod}
\usage{
multimod_forecast_plot_mod(input, output, session, map_input, OBS, model_1,
  model_2, model_3, model_4, return_levels = NULL)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{map_input}{}

\item{model_1}{}

\item{model_2}{}

\item{model_3}{}

\item{model_4}{}

\item{return_levels}{}
}
\description{
Shiny server module to do multi-model plots. Needs new parametrization!
}

