% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{multiple_station_map}
\alias{multiple_station_map}
\title{multiple_station_map}
\usage{
multiple_station_map(stations, selected_regine_main = NULL,
  selected_name = NULL, selected_long = NULL, selected_lat = NULL,
  single_poly = FALSE, variable2plot = "none", radius_function = TRUE,
  popups = FALSE)
}
\arguments{
\item{stations}{}

\item{selected_regine_main}{}

\item{selected_name}{}

\item{selected_long}{}

\item{selected_lat}{}

\item{single_poly}{}

\item{variable2plot}{}

\item{radius_function}{}

\item{popups}{}
}
\description{
mapping function
}

