% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{singlemodel_forecast_plot_modUI}
\alias{singlemodel_forecast_plot_modUI}
\title{singlemodel_forecast_plot_modUI}
\usage{
singlemodel_forecast_plot_modUI(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny UI module to plot ...
}

