% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{which_station_in_polygon}
\alias{which_station_in_polygon}
\title{which_station_in_polygon}
\usage{
which_station_in_polygon(stations, map_selection)
}
\arguments{
\item{stations}{}

\item{map_selection}{}
}
\description{
Sub-function use to find which stations are in the polygon the user has drawn
}

